+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
 -   Improve search result filtering logic
 -   Allow pinning shortcuts directly to the launcher home screen
 -   Another workaround for SteamGuard's non-standard OTP format
+-   Allow importing QR code from images
 
 ### Fixed